Static RAG Agent/requirements.txt
sentence-transformers
faiss-cpu
numpy
python-dotenv
google-generativeai